Adjusting theme colors via Customizer
Well, i thought it was not neccessary to only be able to adjust theme
colors via SCSS and/or custom css. This pull request will make it
possible to adjust them via the customizer. Needs enhancements though.

If anyone may jump in: Try to set up Primary hover colour. It will NOT
get displayed on the live site, but only on the customizer live preview.
Any clues why?