Change version of templates to correct version
Set templates to most updated version from Woo, the version they are based on;
Adjust formatting to be identical to Woo;

Not all versions are 3.3.0, we should reflect the WooCommerce templates so that we know which version they are based on. There are instances  where seemingly older versions are the most current.